apps/simulation/requirements.txt
langchain
langgraph
langchain-openai
pydantic
python-dotenv
faker
